pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.3",
    "nltk>=3.9.1",
    "numpy>=2.3.1",
    "pandas>=2.3.1",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.4",
    "scikit-learn>=1.7.0",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.46.1",
]
